- Detects maps with special requirements such as Mapping Extensions or any other kind - Detects autogenerated maps from Beat Sage (This may not always work but does most of the time)

+++ BeatSaverMapAnalyzer/Extensions/MapDataExtensions.cs
@@ -41,8 +41,8 @@ namespace RandomSongTournamentAssistant.Extensions
 
         public static string GetInfoText(this BeatmapCharacteristicDifficulty beatmapCharacteristicDifficulty, double bpm)
         {
-            int minutes = beatmapCharacteristicDifficulty.Length / 60;
-            int seconds = beatmapCharacteristicDifficulty.Length - minutes * 60;
+            long minutes = beatmapCharacteristicDifficulty.Length / 60;
+            long seconds = beatmapCharacteristicDifficulty.Length - minutes * 60;
             double notesPerSecond = MapTools.GetNotesPerSecond(bpm, beatmapCharacteristicDifficulty.Duration, beatmapCharacteristicDifficulty.Notes);
             string lengthInfo = minutes + " minutes " + seconds + " seconds ";
